Add oneof feature implementation
Seems to work and mostly complete, tests do work.
Design looks good for me now.

There are two classes `OneOf_` and `OneOfPartField`, later they will be
documented.

See comments in corresponding pr.

NB: here is full list of names of system attributes which _must_ not be
used as name of field in OneOf_ message:

- `which_one_of`
- `__set_value__`
- `__fields__`
- `__parts__`

probably this should better not be used too:

- `__init__`
- `__hash__`
- `__eq__`
- `__repr__`
- `__getattr__`
- `__setattr__`
- `__delattr__`